requirements.txt
furiosa-sdk[full] == 0.10.*
opencv-python == 4.9.*
torch == 1.13.1
torchvision == 0.14.1
uvloop
ultralytics
scikit-learn
facenet-pytorch
jinja2
pytest
pycocotools
